requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
pymysql
python-dotenv